Rename controller & manager packages
